lkl: Make virtio queue size configurable
In below cases, LKL (0xc803 offload, 60000 MTU size)
shows "PKT is likely truncated!" message.
Case1: netperf on the host talks to netserver on the LKL through tap.
Case2: netperf on the host talks to netserver on the host through
bridge, tap, and IP routing on LKL.

Optimal queue size seems to depend on the environment.
So I made virtio queue size configurable (default size 128).
Tested different queue size with the two cases.

Case1: netperf on LKL talks to netserver on the host through tap.

The number of "PKT is likely truncated!" message decreases
with the increase of queue size.
When the queue size is more than 256, LKL does not show the message,
though the result of netperf does not change much.

netperf results (TCP_STREAM, offload=0xc803, MTU=60000):

1) queue size=64    3.86 Gbps
2) queue size=128   11.2 Gbps
3) queue size=256   11.7 Gbps
4) queue size=512   11.4 Gbps
5) queue size=1024  11.7 Gbps

Case2: netperf on the host talks to netserver on the host through
bridge, tap, and IP routing on LKL.

The number of "PKT is likely truncated!" message decreases
with the increase of queue size.
When the queue size is more than 1024, LKL does not show the message,
though the result of netperf does not change much.

netperf results (TCP_STREAM, offload=0xc803, MTU=60000):

1) queue size=64    10.3 Gbps
2) queue size=128   13.5 Gbps
3) queue size=256   13.5 Gbps
4) queue size=512   13.2 Gbps
5) queue size=1024  13.2 Gbps
6) queue size=2048  13.5 Gbps
7) queue size 4096  13.4 Gbps

Signed-off-by: Motomu Utsumi <motomuman@gmail.com>