lkl: Separate pipe network backend codes from other virtio devices
This change separates network functions for pipe network backend
from other network backends which use file descriptors.

This helps to introduce specific optimizations for pipe network backend.
pipe specific optimizations are necessary because pipe backend works
differently from other network backends as follows.

When LKL tries to send a large packet to pipe device,
it is possible to write part of the packet.

Moreover, when LKL tries to receive a packet from pipe device,
LKL is not able to detect the border of the packet and
it is possible to read part of the packet or more than one packet.

This behavior is different from other network backends,
and this could lead packet loss and performance reduction.

Signed-off-by: Motomu Utsumi <motomuman@gmail.com>