lkl: Add packet header for pipe network backend
When LKL tries to a large packet to pipe device,
it is possible to write part of the packet.

Moreover, when LKL tries to receive a packet from pipe device,
LKL is not able to detect the border of the packet and
it is possible to read part of the packet or more than one packet.

This behavior is different from other network backends,
and this could lead packet loss and performance reduction.

This commit fixes this issue by adding packet length information
as a header.

Tested with netperf TCP_STREAM.
netperf on LKL talks to netserver on LKL through named pipe.

Before this commit:
1)OFFLOAD=0x0000 (no offload)              MTU  1500  0.018 Gbps
2)OFFLOAD=0x883  (host + guest csum, tso4) MTU 60000  7.1 Gbps

After this commit:
1)OFFLOAD=0x0000 (no offload)              MTU  1500 1.8 Gbps
2)OFFLOAD=0x883  (host + guest csum, tso4) MTU 60000 9.0 Gbps

Signed-off-by: Motomu Utsumi <motomuman@gmail.com>